pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repralign"
version = "0.1.0"
description = "Reusable layer-wise representation alignment analysis for multimodal models."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "matplotlib>=3.7",
  "numpy>=1.24",
  "PyYAML>=6.0",
  "torch>=2.1",
]

[project.optional-dependencies]
huggingface = [
  "transformers>=4.40",
]
dev = [
  "pytest>=8.0",
]

[project.scripts]
repralign = "repralign.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
